+++ lib/wikidosia.js
@@ -1,6 +1,7 @@
 var request = require("request");
 var Twit = require('twit');
 var schedule = require('node-schedule');
+var Promise = require('promise');
 var Wikipedia = require("./wikipedia");
 
 var wikipedia = new Wikipedia();
@@ -125,6 +126,72 @@ Wikidosia.prototype.txiokatuWikidosia = function() {
         });
 
     });
+
+};
+
+Wikidosia.prototype.eskuratuArtikuluenIkustaldienHistoria = function(artikuluak) {
+
+    var promise = new Promise(function(resolve, reject) {
+
+        // Wikipediako artikuluen datuak eskuratzeko egingo ditugun deien promise-ak gordetzeko array-a.
+        var promiseak = [];
+
+        // Datuak array-ko lehen elementuan daten array-a gordeko dugu.
+        // Ondorengo elementuetan artikulu bakoitzaren ikustaldien datuak.
+        var datuak = [];
+
+        // Daten array-a gehitu lehen elementu bezala array-aren izena, "data", duelarik.
+        datuak[0] = ["data"];
+
+        // Artikulu bakoitzeko lehen elementu bezala artikuluaren izena duen array bat gehitu datuen array-ra.
+        artikuluak.forEach(function(artikulua) {
+            datuak.push([artikulua.replace(/_/g, " ")]);
+        });
+
+        // Artikulu bakoitzaren datuak eskuratuko ditugu Wikipediaren APIa erabiliz.
+        // Wikipedia objektuaren getDailyPageViews metodoak itzultzen dituen promise-ak array batean gordeko ditugu.
+        artikuluak.forEach(function(artikulua) {
+
+            promiseak.push(wikipedia.getDailyPageViews("eu", artikulua, "2015", "12", "01", "2016", "01", "05", "all-access", "all-agents"));
+
+        });
+
+        // Datu guztiak eskuratu ditugunean...
+        Promise.all(promiseak).then(function(emaitzak) {
+
+            // Artikulu bakoitzaren datuekin...
+            emaitzak.forEach(function(emaitza, indizea) {
+
+                // Egun bakoitzeko datuekin (data eta ikustaldiak)...
+                emaitza.forEach(function(datua) {
+
+                    // Lehen artikulua bada, egun guztien datak gordeko ditugu daten array-an.
+                    if (indizea === 0) {
+
+                        datuak[0].push(datua.timestamp);
+
+                    }
+
+                    // Artikuluaren ikustaldiak gorde dagokion array-an.
+                    datuak[indizea + 1].push(datua.views);
+
+                });
+
+            });
+
+            resolve(datuak);
+
+        }, function(errorea) {
+
+            console.log(errorea);
+
+            reject(errorea);
+
+        });
+
+    });
+
+    return promise;
 
 };
